+++ R/package-maotai.R
@@ -4,6 +4,7 @@
 #' This package contains following functions, 
 #' \tabular{ll}{
 #' FUNCTION \tab DESCRIPTION \cr
+#' \code{\link{lgpa}} \tab Large-scale Generalized Procrustes Analysis \cr
 #' \code{\link{lyapunov}} \tab Solve Lyapunov Equation \cr
 #' \code{\link{matderiv}} \tab Numerical Approximation to Gradient of a Function with Matrix Argument \cr
 #' \code{\link{pdeterminant}} \tab Calculate the Pseudo-Determinant of a Matrix \cr
@@ -17,6 +18,7 @@
 #' @aliases maotai-package
 #' @import Rdpack
 #' @importFrom stats rnorm
+#' @importFrom shapes procGPA
 #' @importFrom utils packageVersion
 #' @importFrom RSpectra eigs
 #' @importFrom Matrix rankMatrix