+++ NEWS.md
@@ -1,10 +1,11 @@
 # News for Package maotai
-### Changes in version 0.1.0
+### Version 0.1.0
   * Package is first deployed.
   * Initialize the following documentation:
     - NEWS for keeping record of updates.
     - README to briefly introduce the method.
-  * initial set of functions are listed as follows:
+  * Initial set of functions are listed as follows:
+    - `lgpa`         : large-scale generalized Procrustes analysis.
     - `lyapunov`     : solve Lyapunov equation.
     - `matderiv`     : numerical approximation to gradient of a function with matrix argument.
     - `pdeterminant` : calculate the pseudo-determinant of a matrix.